Day 6 (a) passes
Look ma, my first Monad Transformer Stack! Yes it's a crazy mess, and no
I don't quite understand how it all snaps together but it is a
ridiculous feeling when you just plug away at a hacky first pass,
iterate on it until it compiles and BOOM it just works.